Adjusts after DB changes (#5)

+++ src/main/java/com/auxby/usermanager/api/v1/user/UserRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<UserDetails, Long> {
+public interface UserRepository extends JpaRepository<UserDetails, Integer> {
     @EntityGraph(attributePaths = {"addresses", "contacts"})
     Optional<UserDetails> findUserDetailsByUserName(String userName);
 }
